# Write out the loop for better perf on chooseV