@nordfjord 2022-11-12
The toArray was unused as well so I've removed the entire ValueOption module